docs: clarify slidesToScroll=auto behaviour
I spent some time trying different approaches to make the carousel
dynamically scroll according to how many elements are currently visible
in the viewport.

Then I discovered that this option already exists! But the documentation
was not making that as clear as I would have liked.

Here's my attempt at improving the documentation.